% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{genU}
\alias{genU}
\title{Generate the test statistic}
\usage{
genU(st, lab)
}
\arguments{
\item{st}{a ranger forest object.}

\item{lab}{an integer value containing the class labels}
}
\value{
the likelihood-based test statistic
}
\description{
Generate the test statistic
}
